+++ src/lib/compare-deep.ts
@@ -4,7 +4,12 @@ import { match } from './match';
 import { mustGoDeep } from './must-go-deep';
 import { createComparableEntries } from './create-comparable-entries';
 
-export function compareDeep(actual: any, expected: any, tab: string) {
+export function compareDeep(
+  actual: any,
+  expected: any,
+  tab: string,
+  references: Set<unknown>,
+) {
   let deepResult = '';
   const entries = createComparableEntries(actual, expected);
   const prefix = `\n${tab}${lack('...')}\n${tab}`;
@@ -13,11 +18,12 @@ export function compareDeep(actual: any, expected: any, tab: string) {
       if (expectedInfo) {
         const actualValue = actualInfo.value;
         const expectedValue = expectedInfo.value;
-        if (mustGoDeep(actualValue, expectedValue)) {
+        if (mustGoDeep(actualValue, expectedValue, references)) {
           const compareResult = compareDeep(
             actualValue,
             expectedValue,
             `${tab}  `,
+            references,
           );
           if (compareResult) {
             deepResult += `${prefix}${key}: {${compareResult},\n${tab}}`;