+++ test/like.spec.ts
@@ -34,7 +34,43 @@ describe('like', () => {
     expect(error.message).to.match(/Value's not like the expected.+/);
   });
 
+  it('should fail when comparing a non empty object to an empty object', () => {
+    let error: any;
+
+    try {
+      expect({
+        info: 'value',
+      }).to.be.like({});
+    } catch (err) {
+      error = err;
+    }
+
+    expect(error.message).to.match(/Value's not like the expected.+/);
+  });
+
   it('should pass properly comparing a null value with null', () => {
     expect(null).to.have.like(null);
+  });
+
+  it('shoud work fine comparing a recursive object to itself', () => {
+    const test: any = {};
+    test.test = test;
+
+    expect(test).to.be.like(test);
+  });
+
+  it('shoud throw and treated error when comparing a recursive object with something different', () => {
+    let error: any;
+
+    try {
+      const test: any = {};
+      test.test = test;
+
+      expect(test).to.be.like({ test: 'something else' });
+    } catch (err) {
+      error = err;
+    }
+
+    expect(error.message).to.match(/Value's not like the expected.+/);
   });
 });